+++ assets/javascripts/kitten/components/icons/arrow-icon/index.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import styled, { css } from 'styled-components'
 import classNames from 'classnames'
 import PropTypes from 'prop-types'
 import { Deprecated } from '../../../helpers/utils/deprecated'
@@ -10,6 +9,7 @@ export const DeprecatedArrowIconSvg = ({
   direction,
   disabled,
   version,
+  title,
   ...others
 }) => {
   const arrowIconClassNames = classNames(
@@ -28,7 +28,7 @@ export const DeprecatedArrowIconSvg = ({
       className={arrowIconClassNames}
       viewBox="0 0 6 6"
     >
-      <title>Arrow</title>
+      {title && <title>{title}</title>}
       <path d="M6 0H0v6h2V2h4z" />
     </svg>
   )
@@ -72,51 +72,44 @@ DeprecatedArrowIcon.defaultProps = {
   className: '',
 }
 
-const SvgArrow = styled(({ direction, disabled, ...props }) => (
-  <svg {...props} />
-))`
-  transform: ${({ direction }) => {
-    switch (direction) {
-      case 'right':
-        return 'rotate(90deg)'
-      case 'left':
-        return 'rotate(-90deg)'
-      case 'bottom':
-        return 'rotate(180deg)'
-      default:
-        return 'none'
-    }
-  }};
-
-  ${({ disabled }) =>
-    disabled &&
-    css`
-      fill: ${COLORS.background1};
-    `}
-`
-
-export const ArrowIcon = ({ version, direction, disabled, ...others }) => {
+export const ArrowIcon = ({
+  version,
+  direction,
+  disabled,
+  color,
+  title,
+  ...others
+}) => {
   if (version === 'deprecated-center-of-gravity') {
     return (
       <DeprecatedArrowIcon
         version={version}
         direction={direction}
         disabled={disabled}
+        title={title}
         {...others}
       />
     )
   }
 
+  const transform = {
+    right: 'rotate(90)',
+    left: 'rotate(-90)',
+    bottom: 'rotate(180)',
+    top: null,
+  }
+
   return (
-    <SvgArrow
+    <svg
       {...others}
       xmlns="http://www.w3.org/2000/svg"
       viewBox="0 0 8.48 5.64"
-      direction={direction}
-      disabled={disabled}
+      fill={disabled ? COLORS.background1 : color}
+      transform={direction ? transform[direction] : null}
     >
+      {title && <title>{title}</title>}
       <path d="M0 4.24 L4.24,0 L8.48,4.24 L7.08,5.64 L4.24,2.77 L1.4,5.6 z" />
-    </SvgArrow>
+    </svg>
   )
 }
 
@@ -127,6 +120,8 @@ ArrowIcon.propTypes = {
   className: PropTypes.string,
   width: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
   height: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+  color: PropTypes.string,
+  title: PropTypes.string,
 }
 
 ArrowIcon.defaultProps = {
@@ -136,4 +131,6 @@ ArrowIcon.defaultProps = {
   className: '',
   width: '8.48',
   height: '5.64',
+  color: COLORS.font1,
+  title: '',
 }