+++ assets/javascripts/kitten/components/icons/burger-icon/burger-icons.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { boolean, select } from '@storybook/addon-knobs'
+import { boolean, color } from '@storybook/addon-knobs'
 import { BurgerIcon } from './index'
 import { Marger } from '../../layout/marger'
 import { Container } from '../../grid/container'
@@ -23,16 +23,8 @@ export const Default = () => {
     <StoryGrid>
       <BurgerIcon
         isActive={boolean('isActive ?', false)}
-        mainColor={select(
-          'mainColor',
-          { default: COLORS.font1, red: '#FF0000' },
-          COLORS.font1,
-        )}
-        hoverColor={select(
-          'hoverColor',
-          { default: COLORS.primary1, red: '#FF0000' },
-          COLORS.primary1,
-        )}
+        mainColor={color('mainColor', COLORS.font1)}
+        hoverColor={color('hoverColor', COLORS.primary1)}
       />
     </StoryGrid>
   )