+++ assets/javascripts/kitten/components/icons/waiting-icon/index.js
@@ -27,5 +27,5 @@ WaitingIcon.propTypes = {
 
 WaitingIcon.defaultProps = {
   color: '#fff',
-  title: 'Waiting',
+  title: '',
 }